server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,10 @@ const NotFoundError = require("../../exceptions/NotFoundError");
 const { mapDbToModel } = require("../../utils");
 
 class NotesService {
-    constructor(collaborationService) {
+    constructor(collaborationService, cacheService) {
         this._pool = new Pool();
         this._collaborationService = collaborationService;
+        this._cacheService =  cacheService;
     }
 
     async addNote({ title, body, tags, owner }) {
@@ -26,21 +27,28 @@ class NotesService {
             throw new InvariantError('Catatan gagal ditambahkan');
         }
 
+        this._cacheService.delete(`notes:${owner}`);
+
         return result.rows[0].id;
     }
 
     async getNotes(owner) {
-        const query = {
-            text: `SELECT notes.* FROM notes
-            LEFT JOIN collaborations ON collaborations.note_id = notes.id
-            WHERE notes.owner = $1 OR collaborations.user_id = $1
-            GROUP by notes.id`,
-            values: [owner],
-        };
-
-        const result = await this._pool.query(query);
-
-        return result.rows.map(mapDbToModel);
+        try { 
+            const result = await this._cacheService.get(`notes:${owner}`);
+            return JSON.parse(result);
+        } catch(error) {
+            const query = {
+                text: `SELECT notes.* FROM notes
+                LEFT JOIN collaborations ON collaborations.note_id = notes.id
+                WHERE notes.owner = $1 OR collaborations.user_id = $1
+                GROUP by notes.id`,
+                values: [owner],
+            };
+    
+            const result = await this._pool.query(query);
+    
+            return result.rows.map(mapDbToModel);
+        }
     }
 
     async getNoteById(id) {
@@ -64,7 +72,7 @@ class NotesService {
     async editNoteById(id, { title, body, tags }) {
         const updatedAt = new Date().toISOString();
         const query = {
-            text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+            text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
             values: [title, body, tags, updatedAt, id],
         };
 
@@ -73,11 +81,14 @@ class NotesService {
         if (!result.rows.length) {
             throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
         }
+        
+        const {owner} = result.rows[0];
+        this._cacheService.delete(`notes:${owner}`);
     }
 
     async deleteNoteById(id) {
         const query = {
-            text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+            text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
             values: [id],
         };
 
@@ -87,6 +98,8 @@ class NotesService {
             throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
         }
 
+        const {owner} = result.rows[0];
+        this._cacheService.delete(`notes:${owner}`);
     }
 
     async verifyNoteOwner(id, owner) {
